requirements.txt
fastapi
uvicorn
httpx
tenacity
python-dotenv
chromadb
sqlalchemy>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4      